arch_x86: fixed MOV instruction decoding

+++ src/arch/x86/x86_operand.cpp
@@ -179,7 +179,7 @@ static Expression::SPType __DecodeModRmAddress16(CpuInformation* pCpuInfo, Binar
       u8 Disp8;
       rBinStrm.Read(Offset + 1, Disp8);
       OprdLen += sizeof(Disp8);
-      return Expr::MakeBinOp(OperationExpression::OpAdd, spOff, Expr::MakeConst(16, SignExtend<s16, 8>(Disp8)));
+      return Expr::MakeBinOp(OperationExpression::OpAdd, spOff, Expr::MakeConst(16, static_cast<u16>(SignExtend<s16, 8>(Disp8))));
       break;
     }
   case 0x2:
@@ -236,7 +236,7 @@ static Expression::SPType __DecodeSib32(CpuInformation* pCpuInfo, BinaryStream c
 
         u8 Disp8;
         rBinStrm.Read(Offset + sizeof(ModRm), Disp8);
-        spDisp = Expr::MakeConst(32, SignExtend<s32, 8>(Disp8));
+        spDisp = Expr::MakeConst(32, static_cast<u32>(SignExtend<s32, 8>(Disp8)));
         rOprdLen += sizeof(Disp8);
         break;
       }
@@ -258,7 +258,7 @@ static Expression::SPType __DecodeSib32(CpuInformation* pCpuInfo, BinaryStream c
       {
         u8 Disp8;
         rBinStrm.Read(Offset + 1, Disp8);
-        spDisp = Expr::MakeConst(32, SignExtend<s32, 8>(Disp8));
+        spDisp = Expr::MakeConst(32, static_cast<u32>(SignExtend<s32, 8>(Disp8)));
         rOprdLen += sizeof(Disp8);
         break;
       }
@@ -336,7 +336,7 @@ static Expression::SPType __DecodeModRmAddress32(CpuInformation* pCpuInfo, Binar
     {
       u8 Disp8;
       rBinStrm.Read(Offset + sizeof(ModRm), Disp8);
-      spDisp = Expr::MakeConst(32, SignExtend<s32, 8>(Disp8));
+      spDisp = Expr::MakeConst(32, static_cast<u32>(SignExtend<s32, 8>(Disp8)));
       rOprdLen += sizeof(Disp8);
       break;
     }
@@ -384,7 +384,7 @@ static Expression::SPType __DecodeSib64(CpuInformation* pCpuInfo, BinaryStream c
       {
         u32 Disp32;
         rBinStrm.Read(Offset + sizeof(ModRm), Disp32);
-        spDisp = Expr::MakeConst(64, SignExtend<s64, 32>(Disp32));
+        spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 32>(Disp32)));
         rOprdLen += sizeof(Disp32);
         break;
       }
@@ -395,7 +395,7 @@ static Expression::SPType __DecodeSib64(CpuInformation* pCpuInfo, BinaryStream c
 
         u8 Disp8;
         rBinStrm.Read(Offset + sizeof(ModRm), Disp8);
-        spDisp = Expr::MakeConst(64, SignExtend<s64, 8>(Disp8));
+        spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 8>(Disp8)));
         rOprdLen += sizeof(Disp8);
         break;
       }
@@ -417,7 +417,7 @@ static Expression::SPType __DecodeSib64(CpuInformation* pCpuInfo, BinaryStream c
       {
         u8 Disp8;
         rBinStrm.Read(Offset + 1, Disp8);
-        spDisp = Expr::MakeConst(64, SignExtend<s64, 8>(Disp8));
+        spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 8>(Disp8)));
         rOprdLen += sizeof(Disp8);
         break;
       }
@@ -426,7 +426,7 @@ static Expression::SPType __DecodeSib64(CpuInformation* pCpuInfo, BinaryStream c
       {
         u32 Disp32;
         rBinStrm.Read(Offset + 1, Disp32);
-        spDisp = Expr::MakeConst(64, SignExtend<s64, 32>(Disp32));
+        spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 32>(Disp32)));
         rOprdLen += sizeof(Disp32);
         break;
       }
@@ -496,7 +496,7 @@ static Expression::SPType __DecodeModRmAddress64(CpuInformation* pCpuInfo, Binar
     {
       u8 Disp8;
       rBinStrm.Read(Offset + sizeof(ModRm), Disp8);
-      spDisp = Expr::MakeConst(64, SignExtend<s64, 8>(Disp8));
+      spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 8>(Disp8)));
       rOprdLen += sizeof(Disp8);
       break;
     }
@@ -504,7 +504,7 @@ static Expression::SPType __DecodeModRmAddress64(CpuInformation* pCpuInfo, Binar
     {
       u32 Disp32;
       rBinStrm.Read(Offset + sizeof(ModRm), Disp32);
-      spDisp = Expr::MakeConst(64, SignExtend<s64, 32>(Disp32));
+      spDisp = Expr::MakeConst(64, static_cast<u64>(SignExtend<s64, 32>(Disp32)));
       rOprdLen += sizeof(Disp32);
       break;
     }
